pyproject.toml
[project]
name = "pve-vdi-client"
version = "0.0.1"
dependencies = [
    "requests==2.27.1",
    "proxmoxer==1.3.0",
    "python-decouple==3.6",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]